+++ src/settingClasses/ReflowCurveSettings.h
@@ -5,15 +5,15 @@ class ReflowCurveSettings
 {
   public:
     ReflowCurveSettings();
-    void SetPIDTunings(int Kp, int Ki, int Kd);
-    void SetSoak(int RtsTemPerSec, int SoakTime, int SoakTemp);
-    void SetReflow(int RtpTempPerSec, int ReflowMaxTemp, int ReflowTime);
-    void SetCooling(int CoolingTempPerSec);
+    void SetPIDTunings(double Kp, double Ki, double Kd);
+    void SetSoak(double RtsTemPerSec, double SoakTime, double SoakTemp);
+    void SetReflow(double RtpTempPerSec, double ReflowMaxTemp, double ReflowTime);
+    void SetCooling(double CoolingTempPerSec);
 
   //private:
-    int kp, ki, kd;
-    int rtsTempPerSec, soakTemp, rtpTempPerSec, reflowMaxTemp, coolingTempPerSec;
-    int soakTime, reflowTime;
+    double kp, ki, kd;
+    double rtsTempPerSec, soakTemp, rtpTempPerSec, reflowMaxTemp, coolingTempPerSec;
+    double soakTime, reflowTime;
 };
 
 #endif
